backend: finish test school migration data, small fix of TeacherProjection (#34)

+++ backend/src/main/java/ip91/chui/oleh/model/dto/projection/TeacherProjection.java
@@ -3,6 +3,6 @@ package ip91.chui.oleh.model.dto.projection;
 public interface TeacherProjection {
 
   Long getId();
-  int getTotalHours();
+  double getTotalHours();
 
 }
